Shrink mt_evtMenu 00375f30 loop widths

+++ src/Graphics/Model/mdlFile.c
@@ -773,7 +773,7 @@ void FUN_0034d150(int param_1,u32 *param_2);
 u64 FUN_0034d3a0(int param_1,int param_2,int param_3);
 u64 FUN_0034d510(u64 param_1);
 void FUN_0034d670(u64 param_1);
-u64 FUN_0034d6f0(int param_1);
+u32 FUN_0034d6f0(int param_1);
 void FUN_0034d810(int param_1,int param_2);
 void FUN_0034d8a0(int param_1,u16 param_2,int param_3);
 void FUN_0034d990(int param_1,u32 param_2,u32 param_3);
@@ -49839,7 +49839,7 @@ void FUN_0034d670(u64 param_1)
 // FUN_0034D6F0 NONMATCHING
 
 
-u64 FUN_0034d6f0(int param_1)
+u32 FUN_0034d6f0(int param_1)
 {
   int iVar1;
   u32 uVar2;
